fastapi-chat-ws/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
pyjwt==2.9.0
passlib[bcrypt]==1.7.4
pytest==8.2.0
httpx==0.27.2
